+++ packages/walletconnect/src/errors/exports.ts
@@ -1,8 +1,10 @@
 export {
   ClientDisposedError,
   InvalidUriError,
+  PairingCleanupError,
   PairingInProgressError,
   PairingResetError,
+  PairingStartError,
   PairingTimeoutError,
   ProjectIdRequiredError,
   ProposalSettledError,